home coding v2.1

+++ ATOMPSGod/ATOMPSGod/ViewModel/person/ATOMCollectionViewModel.h
@@ -11,7 +11,7 @@
 
 @interface ATOMCollectionViewModel : NSObject
 @property (nonatomic, assign) NSInteger uid;
-@property (nonatomic, assign) int type;
+@property (nonatomic, assign) NSInteger type;
 @property (nonatomic, copy) NSString *userName;
 @property (nonatomic, copy) NSString *userSex;
 @property (nonatomic, copy) NSString *avatarURL;
